dt_gen_ai_analytics-helper/pyproject.toml
[tool.poetry]
name = "dt_gen_ai_analytics_helper"
version = "0.1.0"
description = ""
authors = ["robertanwokonko <roberta.nwokonko@datatonic.com>"]
packages = [{include = "dt_gen_ai_analytics_helper"}]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
langchain = "0.0.240"
google-cloud-aiplatform = "^1.26.0"
google-cloud-storage = "^2.9.0"
google-cloud-bigquery-storage = "^2.22.0"
pydantic = "^1.10.9"
tabulate = "^0.9.0"
torch = "^2.0.1"
grpcio = "^1.54.2"
grpcio-status = "^1.54.2"
uvicorn = "^0.22.0"
fastapi = "^0.98.0"
pandas = "^2.0.2"
pyarrow = "^12.0.1"
sqlalchemy-bigquery = "^1.8.0"
python-dotenv = "^1.0.0"
black = "^23.3.0"
httpx = "^0.24.1"
pytest = "^7.4.0"
gradio = "^3.35.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"
notebook = "^7.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
